refactor: move logstream API
api/rest --> pkg/server
